MDEV-24792 Assertion `!newest_lsn || fil_page_get_type(page)' failed upon MariaBackup prepare in buf_flush_init_for_writing with innodb_log_optimize_ddl=off
fsp_free_page() writes MLOG_INIT_FREE_PAGE, but does not update page
type. But fil_crypt_rotate_page() checks the type to understand if the
page is freshly initialized, and writes dummy record(updates space id)
to force rotation during recovery. This dummy record causes assertion
crash when the page is flushed after recovery, as it's supposed
that pages LSN is 0 for freshly initialized pages.

The fix is to mark page as freed and not to write dummy record during
keys rotation procedure for such marked pages.

The fix is very similiar to MDEV-24695, but as 10.5 code is quite
different, MDEV-24695 can not be backported as is. There is no need to
port the current fix to 10.5+, as MDEV-24695 is already fixed for 10.5+.

This commit does not contain test case yet.