MDEV-24792 Assertion `!newest_lsn || fil_page_get_type(page)' failed upon MariaBackup prepare in buf_flush_init_for_writing with innodb_log_optimize_ddl=off
Addressed code review notes and added test case.

The test case is almost the same as for MDEV-24695. The general difference
between 10.4 and 10.5 is that in 10.5 the assertion crashes during redo
log appying, while in 10.4 it crashes during page flushing. So to make
it crashed page flushing is forced after each server restart in the
test.